Crate wifiscanner

A crate to list WiFi hotspots in your area.

As of v0.3.x now supports OSX and Linux. Windows to follow.

Usage

This crate is on crates.io and can be used by adding wifiscanner to the dependencies in your project's Cargo.toml.

[dependencies]
wifiscanner = "0.3.*"

and this to your crate root:

extern crate wifiscanner;

Example

use wifiscanner;
println!("{:?}", wifiscanner::scan());

Alternatively if you've cloned the the Git repo, you can run the above example using: cargo run --example scan.

Structs

Wifi
Wifi struct used to return information about wifi hotspots

Enums

Error

Functions

scan
Returns a list of WiFi hotspots in your area - (Linux) uses iwlist